+++ js/multimodaltextbox/shared/MultimodalTextbox.svelte
@@ -68,6 +68,7 @@
 
 	$: if (value === null) value = { text: "", files: [] };
 	$: value, el && lines !== max_lines && resize(el, lines, max_lines);
+	$: can_submit = value.text !== "";
 
 	const dispatch = createEventDispatcher<{
 		change: typeof value;
@@ -127,7 +128,9 @@
 		await tick();
 		if (e.key === "Enter" && e.shiftKey && lines > 1) {
 			e.preventDefault();
-			dispatch("submit");
+			if (can_submit) {
+				dispatch("submit");
+			}
 		} else if (
 			e.key === "Enter" &&
 			!e.shiftKey &&
@@ -135,7 +138,9 @@
 			max_lines >= 1
 		) {
 			e.preventDefault();
-			dispatch("submit");
+			if (can_submit) {
+				dispatch("submit");
+			}
 		}
 	}
 
@@ -247,7 +252,7 @@
 >
 	<!-- svelte-ignore a11y-autofocus -->
 	<label class:container>
-		<BlockTitle {show_label} {info}>{label}</BlockTitle>
+		<BlockTitle {root} {show_label} {info}>{label}</BlockTitle>
 		{#if value.files.length > 0 || uploading}
 			<div
 				class="thumbnails scroll-hide"
@@ -341,6 +346,7 @@
 					class="submit-button"
 					class:padded-button={submit_btn !== true}
 					on:click={handle_submit}
+					disabled={!can_submit}
 				>
 					{#if submit_btn === true}
 						<Send />
@@ -449,6 +455,13 @@
 	.upload-button:hover,
 	.submit-button:hover {
 		background: var(--button-secondary-background-fill-hover);
+	}
+
+	.stop-button:disabled,
+	.upload-button:disabled,
+	.submit-button:disabled {
+		background: var(--button-secondary-background-fill);
+		cursor: initial;
 	}
 
 	.submit-button :global(svg) {