+++ js/simpledropdown/Index.svelte
@@ -15,6 +15,7 @@
 	export let scale: number | null = null;
 	export let min_width: number | undefined = undefined;
 	export let loading_status: LoadingStatus;
+	export let root: string;
 	export let gradio: Gradio<{
 		change: string;
 		input: never;
@@ -62,7 +63,7 @@
 	{/if}
 
 	<label class:container>
-		<BlockTitle {show_label} info={undefined}>{label}</BlockTitle>
+		<BlockTitle {root} {show_label} info={undefined}>{label}</BlockTitle>
 		<select disabled={!interactive} bind:value={display_value}>
 			{#each choices as choice}
 				<option>{choice[0]}</option>