+++ js/dataframe/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @gradio/dataframe
 
+## 0.11.0-beta.7
+
+### Dependency updates
+
+- @gradio/statustracker@0.8.0-beta.4
+- @gradio/atoms@0.9.0-beta.4
+- @gradio/client@1.6.0-beta.4
+- @gradio/upload@0.13.0-beta.6
+- @gradio/markdown@0.10.0-beta.4
+- @gradio/button@0.3.0-beta.6
+
 ## 0.11.0-beta.6
 
 ### Features